% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{check_transformation_list}
\alias{check_transformation_list}
\title{check does the transformation which will be implemented already implemented}
\usage{
check_transformation_list(trans_to_do, trans)
}
\arguments{
\item{trans_to_do}{list of transformation to implementation}

\item{trans}{list of transformations which already implemented}
}
\value{

}
\description{
check does the transformation which will be implemented already implemented
}
