% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{cluster_merging}
\alias{cluster_merging}
\title{Merging two or more clusters within cluster annotating vector}
\usage{
cluster_merging(clusters, cluster_to_merge)
}
\arguments{
\item{clusters}{vector of clusters}

\item{cluster_to_merge}{vector of cluster name to merging}
}
\value{

}
\description{
Merging two or more clusters within cluster annotating vector
}
