% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{cytofAsinh}
\alias{cytofAsinh}
\title{CytofAsinh fron cytofkit articles}
\usage{
cytofAsinh(value, cofactor = 5)
}
\arguments{
\item{value}{value or vector to transform}

\item{cofactor}{number to divide}
}
\value{

}
\description{
CytofAsinh fron cytofkit articles
}
