% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{cytofLog10}
\alias{cytofLog10}
\title{cytofLog10 from cytofkit articles}
\usage{
cytofLog10(value)
}
\arguments{
\item{value}{value or vector to transform}
}
\value{

}
\description{
cytofLog10 from cytofkit articles
}
