% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{exprs_asinh_transformation}
\alias{exprs_asinh_transformation}
\title{Transformation exprs matrix by asinh fransformation}
\usage{
exprs_asinh_transformation(exprs_data, use_markers, cofactor = 5)
}
\arguments{
\item{exprs_data}{matrix with expression data}

\item{cofactor}{digit with used as denominator to transform data}

\item{use_marker}{list of marker for transformation}
}
\value{
flowSet object with transform data by asinh function and divided by cofactor
}
\description{
Transformation exprs matrix by asinh fransformation
}
