% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{filter_edges}
\alias{filter_edges}
\title{Filter out edges which have weight more than threshold}
\usage{
filter_edges(edges, threshold)
}
\arguments{
\item{edges}{edges created by distance between clusters}

\item{threshold}{length of edges to filtering out}
}
\value{

}
\description{
Filter out edges which have weight more than threshold
}
