% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{get_abundance_dataframe}
\alias{get_abundance_dataframe}
\title{Create the data table to draw the abundance barplot}
\usage{
get_abundance_dataframe(fcs_raw, cell_clustering)
}
\arguments{
\item{fcs_raw}{fcs object}

\item{cell_clustering}{vector of marker to clustering}
}
\value{

}
\description{
Create the data table to draw the abundance barplot
}
