% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gating.R
\name{get_add_cell_annotation_from_data}
\alias{get_add_cell_annotation_from_data}
\title{Extract annotation fron fcs_raw object to cell annotation object}
\usage{
get_add_cell_annotation_from_data(entire_panel, fcs_raw, cell_annotation)
}
\arguments{
\item{entire_panel}{wide range of column names from floeSet without any filtration}

\item{fcs_raw}{flowSet object with data}

\item{cell_annotation}{cell annotation data frame to save new annotation from gates}
}
\value{

}
\description{
Extract annotation fron fcs_raw object to cell annotation object
}
