% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{get_all_consensusClust}
\alias{get_all_consensusClust}
\title{Extration clustering information for all calculated k in ConsensusClustPluse}
\usage{
get_all_consensusClust(som, mc)
}
\arguments{
\item{mc}{ConsensusClusterPlus object}
}
\value{
matrix with row per cell and colomn for each clustering run of ConsensusCluster
}
\description{
Extration clustering information for all calculated k in ConsensusClustPluse
}
