% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{get_ann_adding_fcs_files}
\alias{get_ann_adding_fcs_files}
\title{Adding of annotation info to flowSet object}
\usage{
get_ann_adding_fcs_files(fcs_raw, cell_ann, col_name = NULL)
}
\arguments{
\item{fcs_raw}{flowSet object}

\item{cell_ann}{cell annotation table}

\item{col_name}{vetor of column names from cell annotation data which we would like to add to fcs}
}
\value{

}
\description{
Adding of annotation info to flowSet object
}
