% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gating.R
\name{get_cell_annotation_fcs_files}
\alias{get_cell_annotation_fcs_files}
\title{Adding of cluster-info to flowSet object}
\usage{
get_cell_annotation_fcs_files(fcs_raw, cell_annotation, column_names = NULL)
}
\arguments{
\item{fcs_raw}{flowSet object with data}

\item{cell_clustering_list}{vector of cluster annotation for cells in order of expression data}
}
\value{

}
\description{
Adding of cluster-info to flowSet object
}
