% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gating.R
\name{get_cell_type_from_gates}
\alias{get_cell_type_from_gates}
\title{Function adds set of gates to cell annotation object}
\usage{
get_cell_type_from_gates(
  gates,
  gate_list,
  cell_annotation,
  method = "pure",
  name_new_ann = NULL
)
}
\arguments{
\item{gates}{data frame with allocated gates with logical values in it}

\item{gate_list}{name of the gates, column names from gates data frame}

\item{cell_annotation}{cell annotation data frame to save new annotation from gates}

\item{method}{methd of converting gates to cell annotation "pure" or "squeeze"}
}
\value{

}
\description{
Function adds set of gates to cell annotation object
}
