% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{get_clustered_fcs_files}
\alias{get_clustered_fcs_files}
\title{Adding of cluster-info to flowSet object}
\usage{
get_clustered_fcs_files(fcs_raw, cell_clustering_list, column_name = "cluster")
}
\arguments{
\item{fcs_raw}{fcs object}

\item{cell_clustering_list}{list of clusters for ech samples}
}
\value{

}
\description{
Adding of cluster-info to flowSet object
}
