% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_panel_analysis.R
\name{get_clustering_exp_cell_f_cross_panel}
\alias{get_clustering_exp_cell_f_cross_panel}
\title{Get dxpressing cell fraction data for target marker}
\usage{
get_clustering_exp_cell_f_cross_panel(
  fcs_raw_set,
  cell_clustering_list_set,
  use_samples,
  use_marker_set,
  target_mk,
  min_cell_number = 5
)
}
\arguments{
\item{fcs_raw_set}{list of flowSet objects}

\item{cell_clustering_list_set}{list of vectors with cluster info}

\item{use_samples}{list of used samples}

\item{use_marker_set}{list of used markers}

\item{target_mk}{marker of interest}

\item{min_cell_number}{mininal cell number across the samples}
}
\value{

}
\description{
Get dxpressing cell fraction data for target marker
}
