% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichments.R
\name{get_color_enrichments}
\alias{get_color_enrichments}
\title{Get vector of colors for heatmap}
\usage{
get_color_enrichments(pal = "RdYlBu", pal_reverse = TRUE, range = 100)
}
\arguments{
\item{pal}{name of palette for brewer.pal function (RdYlBu, RdBu, Spectral, PiYG etc.)}

\item{pal_reverse}{reverse or not the order of colors in palette}

\item{range}{number of demanded colours}
}
\value{

}
\description{
Get vector of colors for heatmap
}
