% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{get_consensusClust}
\alias{get_consensusClust}
\title{Get mc consensusCluster object from ConsensusClusterPlus}
\usage{
get_consensusClust(som, maxK = 20)
}
\arguments{
\item{som}{som object from FlowSOM package}

\item{maxK}{maximum number of allocated clusters}
}
\value{
ConsensusClusterPlus object from the same package
}
\description{
Get mc consensusCluster object from ConsensusClusterPlus
}
