% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gating.R
\name{get_data_for_gating}
\alias{get_data_for_gating}
\title{Take data to gating scatterplot}
\usage{
get_data_for_gating(fcs_raw, gating_subset, gating_mk1, gating_mk2)
}
\arguments{
\item{fcs_raw}{flowSet object with data}

\item{gating_subset}{vector of logic values for expression matrix extracted from flowSet object}

\item{gating_mk1}{marker for plotting X axis of density plot for gating}

\item{gating_mk2}{marker for plotting Y axis of density plot for gating}
}
\value{

}
\description{
Take data to gating scatterplot
}
