% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichments.R
\name{get_deconvol_enrichments}
\alias{get_deconvol_enrichments}
\title{get tidy data frame for deconvolution heatmap plot}
\usage{
get_deconvol_enrichments(
  exprs_data,
  mk,
  samples,
  groups,
  summarize_mode = "median"
)
}
\arguments{
\item{exprs_data}{table with enrichment data}

\item{mk}{marker for deconvolution}

\item{samples}{vector with sample information for each cell of enrichment data}

\item{groups}{vector with group information for each cell of enrichment data}

\item{summarize_mode}{method to calculate enrichment of cells groups("median" or "mean")}
}
\value{

}
\description{
get tidy data frame for deconvolution heatmap plot
}
