% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{get_dim_reduce}
\alias{get_dim_reduce}
\title{Preparing data to dimensional reduce plot}
\usage{
get_dim_reduce(
  exprs_data,
  cell_ann,
  subset_coord,
  use_markers,
  force = FALSE,
  method = "tSNE",
  perplexity = 30,
  theta = 0.5,
  max_iter = 1000,
  pca_param = FALSE,
  check_duplicates = FALSE,
  seed = 1234
)
}
\arguments{
\item{exprs_data}{matrix with enrichment data}

\item{cell_ann}{cell annotation data}

\item{subset_coord}{coordinates of subset}

\item{use_markers}{list of markers for dimensional reducing}

\item{force}{flag to make dimensional reducing}

\item{method}{method to dimensional reducing (tSNE or UMAP)}

\item{perplexity}{perplexity for tSNE}

\item{theta}{theta for tSNE}

\item{max_iter}{max iterations for tSNE}

\item{pca_param}{pca parameters for Rtsne function}

\item{check_duplicates}{check_duplicates parameters for Rtsne function}

\item{seed}{seed for randomisation of Rtsne function}
}
\value{

}
\description{
Preparing data to dimensional reduce plot
}
