% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{get_edges}
\alias{get_edges}
\title{Get edges}
\usage{
get_edges(cluster_euclidean_distance)
}
\arguments{
\item{cluster_euclidean_distance}{table with distance between clusters}
}
\value{

}
\description{
Get edges
}
