% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{get_euclid_dist}
\alias{get_euclid_dist}
\title{Get euclidean distance between clones}
\usage{
get_euclid_dist(exprs_data, use_markers, cell_clustering)
}
\arguments{
\item{exprs_data}{matrix with enrichment data}

\item{use_markers}{vector of markers for assay}

\item{cell_clustering}{vector with clusters}
}
\value{

}
\description{
Get euclidean distance between clones
}
