% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{get_exprs_data}
\alias{get_exprs_data}
\title{Get enrichment data from flowFrame object to matrix format}
\usage{
get_exprs_data(fcs_raw)
}
\arguments{
\item{fcs_raw}{flowSet object}
}
\value{

}
\description{
Get enrichment data from flowFrame object to matrix format
}
