% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gating.R
\name{get_exprs_data_for_gating}
\alias{get_exprs_data_for_gating}
\title{extract data table for plot to gate}
\usage{
get_exprs_data_for_gating(exprs_data, gating_subset, gating_mk1, gating_mk2)
}
\arguments{
\item{exprs_data}{matrix with expression data}

\item{gating_subset}{subset to plot to gating}

\item{gating_mk1}{marker for X axis for gating plot}

\item{gating_mk2}{marker for Y axis for gating plot}
}
\value{

}
\description{
extract data table for plot to gate
}
