% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{get_extr_ann}
\alias{get_extr_ann}
\title{Detect columns with annotations}
\usage{
get_extr_ann(exprs_data)
}
\arguments{
\item{fcs_raw}{flowDet object}
}
\value{

}
\description{
Detect columns with annotations
}
