% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{get_extr_transformations}
\alias{get_extr_transformations}
\title{Detecting of implemented transformation by data}
\usage{
get_extr_transformations(exprs_data, use_markers, mode = "dataset")
}
\arguments{
\item{exprs_data}{data frame with enrichment data}

\item{use_markers}{list of markers to consider}

\item{mode}{mode of running the function ("dataset": to consider all data  together, "markers": to consider each marker independently)}
}
\value{

}
\description{
Detecting of implemented transformation by data
}
