% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{get_fcs_cell_clustering_vector}
\alias{get_fcs_cell_clustering_vector}
\title{Create the cluster annotation as one vector for cluster info extraction from fcs file}
\usage{
get_fcs_cell_clustering_vector(cell_clustering_list)
}
\arguments{
\item{cell_clustering_list}{list of clusters for ech samples}
}
\value{

}
\description{
Create the cluster annotation as one vector for cluster info extraction from fcs file
}
