% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{get_fcs_cluster_annotation}
\alias{get_fcs_cluster_annotation}
\title{Create the cluster annotation separated by samples from fcs files by specific coloumn}
\usage{
get_fcs_cluster_annotation(fcs_raw, pattern = "clust")
}
\arguments{
\item{fcs_raw}{fcs object}

\item{pattern}{name or part of colomn name with cluster information. For cytofkit2 output it is "FlowSOM_clusterIDs"}
}
\value{

}
\description{
Create the cluster annotation separated by samples from fcs files by specific coloumn
}
