% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gating.R
\name{get_gates_overlap_vector}
\alias{get_gates_overlap_vector}
\title{Create tidy data frame to order overlaped gates to plotting}
\usage{
get_gates_overlap_vector(gates)
}
\arguments{
\item{gates}{dataframe with gates data}
}
\value{

}
\description{
Create tidy data frame to order overlaped gates to plotting
}
