% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_panel_analysis.R
\name{get_mk_exp_data_cross_p}
\alias{get_mk_exp_data_cross_p}
\title{Title}
\usage{
get_mk_exp_data_cross_p(fcs_raw_set, use_samples, use_marker_set, target_mk)
}
\arguments{
\item{fcs_raw_set}{list of flowSet objects}

\item{use_samples}{list with vectors of samples}

\item{use_marker_set}{list with vectors of markers for analysis}

\item{target_mk}{list of marker of interest}
}
\value{

}
\description{
Title
}
