% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gating.R
\name{get_modif_sparse_data_gating}
\alias{get_modif_sparse_data_gating}
\title{Check and adjusting sparse data for density plot for gating}
\usage{
get_modif_sparse_data_gating(
  gated_data_subset,
  mk = c("gating_mk1", "gating_mk2")
)
}
\arguments{
\item{gated_data_subset}{matrix of data for adjusting}

\item{mk}{column in matrix for adjusting and ploting}
}
\value{

}
\description{
Check and adjusting sparse data for density plot for gating
}
