% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{get_new_fcs_names}
\alias{get_new_fcs_names}
\title{Get new names with additional prefix for samples to save}
\usage{
get_new_fcs_names(old_name, folder_path, prefix = "")
}
\arguments{
\item{old_name}{vector of assumed names to save}

\item{folder_path}{directory to save file}

\item{prefix}{additional prefix to add to names}
}
\value{

}
\description{
Get new names with additional prefix for samples to save
}
