% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{get_nodes}
\alias{get_nodes}
\title{Get nodes}
\usage{
get_nodes(edges, cell_clustering)
}
\arguments{
\item{edges}{edges created by distance between clusters}

\item{cell_clustering}{vector with clusters}
}
\value{

}
\description{
Get nodes
}
