% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{get_optimal_clusters}
\alias{get_optimal_clusters}
\title{Get a optimal number of clusters}
\usage{
get_optimal_clusters(mc, rate_var_expl = 0.9)
}
\arguments{
\item{mc}{ConsensusClusterPlus object}

\item{rate_var_expl}{the portion of variability to explain by clustering}
}
\value{

}
\description{
Get a optimal number of clusters
}
