% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichments.R
\name{get_scale_hm_enrichments}
\alias{get_scale_hm_enrichments}
\title{Scalin data table for enrichments heatmap}
\usage{
get_scale_hm_enrichments(hm_enrich_data, method = "zscore")
}
\arguments{
\item{hm_enrich_data}{table with enrichments data}

\item{method}{method to scale data ("zscore", "norm" or "none")}
}
\value{

}
\description{
Scalin data table for enrichments heatmap
}
