% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gating.R
\name{get_size_subset_gating_data}
\alias{get_size_subset_gating_data}
\title{Subset of data for gating}
\usage{
get_size_subset_gating_data(
  gated_data_subset,
  size_fuse = 10000,
  centers = 100,
  iter.max = 10
)
}
\arguments{
\item{gated_data_subset}{data with 2columns for subset}

\item{size_fuse}{available size of subset}

\item{centers}{number of k for k-mean grouping}

\item{iter.max}{number of iterations for k-mean grouping}
}
\value{

}
\description{
Subset of data for gating
}
