% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{get_subset_coord}
\alias{get_subset_coord}
\title{Subsempling data from exprs data format with sample and fuze considering}
\usage{
get_subset_coord(cell_ann, sampling_size = 0.5, fuse = TRUE, size_fuse = 3000)
}
\arguments{
\item{cell_ann}{cell annotation data}

\item{sampling_size}{fraction of cell amount to plotting}

\item{fuse}{logic value to use fuse for data size}

\item{size_fuse}{number of cells to as max number as fuze for data size}
}
\value{

}
\description{
Subsempling data from exprs data format with sample and fuze considering
}
