% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{get_transformations}
\alias{get_transformations}
\title{Implement the transformation from trans_to_do if they are not in trans list}
\usage{
get_transformations(
  trans_to_do,
  exprs_data,
  use_markers,
  cofactor = 5,
  quantile = 0.01
)
}
\arguments{
\item{trans_to_do}{list of transformations for implementation}

\item{exprs_data}{data frame with enrichment data}

\item{use_markers}{list of considered markers}

\item{cofactor}{co-factior for asinh transformation}

\item{quantile}{quantile for outlier detection}
}
\value{

}
\description{
Implement the transformation from trans_to_do if they are not in trans list
}
