% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{get_UMAP_dataframe}
\alias{get_UMAP_dataframe}
\title{Title}
\usage{
get_UMAP_dataframe(
  fcs_raw,
  use_markers,
  clust_markers,
  tsne_inds,
  cell_clustering,
  method = "UMAP",
  perplexity = 30,
  theta = 0.5,
  max_iter = 1000
)
}
\arguments{
\item{fcs_raw}{fcs object}

\item{use_markers}{vector of marker to extract the data}

\item{clust_markers}{vector of marker to clustering}

\item{tsne_inds}{index to do tSNE}

\item{cell_clustering}{vector of clusters}
}
\value{

}
\description{
Title
}
