% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{outlier_by_quantile_transformation}
\alias{outlier_by_quantile_transformation}
\title{Transformation to a from 0 to 1 variable and removing outliers}
\usage{
outlier_by_quantile_transformation(
  fcs_raw,
  quantile = 0.01,
  use_markers = NULL
)
}
\arguments{
\item{fcs_raw}{flowSet object}

\item{quantile}{number from 0 to 1 to set the quantile which is of cutoff threshold}

\item{use_marker}{list of markers for transformation}
}
\value{

}
\description{
Transformation to a from 0 to 1 variable and removing outliers
}
