% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{scatter_plot_data_prep}
\alias{scatter_plot_data_prep}
\title{Preparing data to tSNE}
\usage{
scatter_plot_data_prep(
  fcs_raw,
  use_markers,
  sampling_size = 0.5,
  method = "tSNE",
  perplexity = 30,
  theta = 0.5,
  max_iter = 1000,
  size_fuse = 5000
)
}
\arguments{
\item{fcs_raw}{flowSet object}

\item{use_markers}{list of markers for tSNE}

\item{sampling_size}{fraction of cell amount for analysis}
}
\value{

}
\description{
Preparing data to tSNE
}
