% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{upload_fcs_data}
\alias{upload_fcs_data}
\title{Upload data from fcs files to write flowSet object, panel and use_marker}
\usage{
upload_fcs_data(fcs_files)
}
\arguments{
\item{fcs_files}{list of path to FCS files}
}
\value{
list with three elements: flowSet object, panel info and use_marker vector
}
\description{
Upload data from fcs files to write flowSet object, panel and use_marker
}
