% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_preparation.R
\name{get_clustered_fcs_files}
\alias{get_clustered_fcs_files}
\title{Adding of annotation info to flowSet object}
\usage{
get_clustered_fcs_files(fcs_raw, cell_ann, column_name = c("cluster"))
}
\arguments{
\item{fcs_raw}{flowSet object}

\item{cell_ann}{cell annotation table}

\item{column_name}{vetor of column names from cell annotation data which we would like to add to fcs}
}
\value{

}
\description{
Adding of annotation info to flowSet object
}
