% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{get_som}
\alias{get_som}
\title{Get som object from flowSOM}
\usage{
get_som(fcs_raw, clust_markers, seed = 1234)
}
\arguments{
\item{fcs_raw}{fcs object}

\item{use_markers}{vector  of markers to clustering}
}
\value{
som object from FlowSOM package
}
\description{
Get som object from flowSOM
}
